Rename the routepolicy package name and update imports
