+++ internal/kgateway/extensions2/plugins/trafficpolicy/ai_policy_test.go
@@ -1,4 +1,4 @@
-package routepolicy
+package trafficpolicy
 
 import (
 	"os"
