+++ internal/kgateway/extensions2/plugins/trafficpolicy/extproc_policy.go
@@ -1,4 +1,4 @@
-package routepolicy
+package trafficpolicy
 
 import (
 	"fmt"
