+++ internal/kgateway/extensions2/plugins/trafficpolicy/extproc_policy_test.go
@@ -1,4 +1,4 @@
-package routepolicy
+package trafficpolicy
 
 import (
 	"testing"
