+++ internal/kgateway/extensions2/plugins/trafficpolicy/local_rate_limit_plugin.go
@@ -1,4 +1,4 @@
-package routepolicy
+package trafficpolicy
 
 import (
 	"time"
