+++ internal/kgateway/extensions2/plugins/trafficpolicy/status.go
@@ -1,4 +1,4 @@
-package routepolicy
+package trafficpolicy
 
 import (
 	"context"
