+++ internal/kgateway/extensions2/plugins/trafficpolicy/traffic_policy_plugin.go
@@ -1,4 +1,4 @@
-package routepolicy
+package trafficpolicy
 
 import (
 	"context"
